:recycle: Merge ProcessInstanceConverter into ProcessModelService
