+++ src/converters/index.ts
@@ -1,6 +1,5 @@
 export * from './empty_activity_converter';
 export * from './event_converter';
-export * from './process_instance_converter';
 export * from './process_model_converter';
 export * from './user_task_converter';
 export * from './manual_task_converter';
